# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DriveTrack - Driving School Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings/change-requests:
    get:
      operationId: listBookingChangeRequests
      tags: [scheduling]
      summary: List booking change requests (school-admin view)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, denied, withdrawn]
      responses:
        "200":
          description: List of change requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingChangeRequest"
  /bookings/{id}/change-requests:
    post:
      operationId: createBookingChangeRequest
      tags: [scheduling]
      summary: Submit a change request for a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestType]
              properties:
                requestType:
                  type: string
                  enum: [cancel, reschedule, availability_override]
                requestedPayloadJson:
                  type: object
      responses:
        "201":
          description: Change request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingChangeRequest"
        "404":
          description: Booking not found
  /bookings/change-requests/{id}/review:
    patch:
      operationId: reviewBookingChangeRequest
      tags: [scheduling]
      summary: Approve or deny a booking change request (school-admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approved, denied]
                reviewNotes:
                  type: string
      responses:
        "200":
          description: Request reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingChangeRequest"
        "404":
          description: Not found
        "409":
          description: Not pending

  # ─── Notifications ────────────────────────────────────────────────────────────
...
  /billing/subscription:
    get:
      operationId: getMySubscription
      tags: [billing]
      summary: Get current subscription and pricing info
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionInfo"
    post:
      operationId: createOrUpdateSubscription
      tags: [billing]
      summary: Create or update a subscription (stub)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planCode]
              properties:
                planCode:
                  type: string
                schoolId:
                  type: integer
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  schemas:
...
    BookingChangeRequest:
      type: object
      required: [id, bookingId, requestedByUserId, requestType, status]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        requestedByUserId:
          type: integer
        requestType:
          type: string
          enum: [cancel, reschedule, update]
        requestedDate:
          type: ["string", "null"]
        requestedTime:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, denied, withdrawn]
        reviewedByUserId:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Notification preferences ─────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [id, userId, planCode, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        schoolId:
          type: ["integer", "null"]
        planCode:
          type: string
        billingProvider:
          type: ["string", "null"]
        externalSubscriptionId:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, trialing, past_due, cancelled, paused]
        currentPeriodStart:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
        seatCount:
          type: ["integer", "null"]
        amountCents:
          type: ["integer", "null"]
        currency:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Feature entitlements ─────────────────────────────────────────────────
...
    SubscriptionInfo:
      type: object
      required: [planCode, status]
      properties:
        planCode:
          type: string
        status:
          type: string
        billingProvider:
          type: ["string", "null"]
        seatCount:
          type: ["integer", "null"]
        renewalAt:
          type: ["string", "null"]
        enforcementEnabled:
          type: boolean
        pricingInfo:
          type: object
          properties:
            viewer:
              type: integer
            independentInstructor:
              type: integer
            schoolBase:
              type: integer
            schoolAdditionalSeat:
              type: integer
...